generic type and constants moved to httpproxy.go
